refactor: move sql generation to helpers
